ClientSpec split
